requirements.txt
python-dotenv
openai
google-generativeai
httpx
pytz
streamlit
pandas
plotly
pytz
requests
